Data Raw Upload + Notebook Creation
With this update we inserted part of the remaining data raw missing: comprising the the regions considered for the database that roughly translate into the insularity and continental regions of each country by continent; and the Taxonomy data that comprise the most accepted name for each species registered.

We also created the jupyter notebook where we will insert all the code used to develop the database.