Taxonomy Update
Taxonomy update search should go through a loop, confirming and correcting data from GBIF and Names verifier. This way the name is at start seached in GBIF, uniformized in GNV and then filtered and reseted on GBIF. This way it is guaranteed that each species if correcly introduced is only kept once.